requirements.txt
# Dependências principais do seu projeto (baseado no pyproject.toml)
google-api-python-client>=2.166.0
google-auth-oauthlib>=1.2.1
langchain-community>=0.3.21
langchain-core>=0.3.51
langchain-experimental>=0.3.4
langchain-groq>=0.3.2
langchain>=0.3.23
langgraph-checkpoint-sqlite>=2.0.6
langgraph>=0.3.27
oauth2client>=4.1.3
python-dotenv>=1.1.0

# Novas dependências para o servidor FastAPI
fastapi
uvicorn[standard]
pydantic
psycopg[binary]
langgraph-checkpoint-pg